% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe.R
\name{epe}
\alias{epe}
\title{Function to calculate the estimated period of exposure to dispensed medicines based on dispensing dates (aka the waiting time distribution)}
\usage{
epe(
  data,
  id,
  pbs_code,
  atc_code,
  date_dispensing,
  use_75th_percentile = TRUE,
  selected_pbs_codes = unique(pbs_code),
  max_time_period = 180
)
}
\arguments{
\item{data}{data.frame containing input data; check your input data is a data.frame and not a tibble, coerce to data.frame if necessary. To use, specify data=your_input_data}

\item{id}{name of character vector in `data` specifying a unique identifier for individual people/patients (can handle integer variables). To use, specify id="your_id_variable_name"}

\item{pbs_code}{name of character vector in `data` specifying the PBS code for the medicine. To use, specify pbs_code="your_pbs_code_variable_name"}

\item{atc_code}{name of character vector in `data` specifying the ATC code for the medicine. To use, specify atc_code="your_atc_code_variable_name"}

\item{date_dispensing}{name of date vector in `data` specifying the dates on which each medicine were dispensed. To use, specify date_dispensing="your_date_of_dispensing_variable_name"}

\item{use_75th_percentile}{logical vector to specify whether the 75th percentile should be used as the EPE value. Default is `TRUE` i.e. the 75th percentile is used. If set to `FALSE` then the median is used instead}

\item{selected_pbs_codes}{optional character vector specifying the PBS codes of interest if EPE is to be calculated for those codes only. If you don't use or specify anything for this argument, the code will still run and use all PBS codes. To use, specify selected_pbs_codes=c("pbs_code1", "pbs_code2", "pbs_code3" etc)}

\item{max_time_period}{integer value specifying the maximum permitted number of days between dispensing to be included in the distribution of time periods used to calculate the EPE. Time periods longer than this value will be excluded. Default value is 180 days i.e. if you don' use or specify anything for this argument, the code will still run and use 180 days as this value}
}
\value{
Object of class list, the first element of which is a data.frame containing the EPE values for each PBS code. The second element is a panel of plots for each PBS code visualising the distributions of time between dispensings (after gaps in treatment were identified and excluded).
}
\description{
Uses dispensing dates to calculate the estimated period of exposure (EPE) as the 75th percentile of the distribution of time periods between dispensings of medicines at the level of PBS codes.
Gaps in dispensing are excluded from this period based on a default of 180 days (although alternative values can be specified).
}
\examples{
epe(data=your_input_data, id="your_id_variable_name", pbs_code="your_pbs_code_variable_name", date_dispensing="your_dispensing_date_variable_name")

epe(data=your_input_data, id="your_id_variable_name", pbs_code="your_pbs_code_variable_name", date_dispensing="your_dispensing_date_variable_name",
selected_pbs_codes = c("code1", "code2", "code3")
)

epe(data=your_input_data, id="your_id_variable_name", pbs_code="your_pbs_code_variable_name", date_dispensing="your_dispensing_date_variable_name",
max_time_period = 120
)

Example using 'mydata' sample dataset provided with the package
epe(data=mydata, id="PPN", pbs_code="PBS", date_dispensing="dispensing_date")

}
\references{
Anton Pottegård and Jesper Hallas, "Assigning exposure duration to single prescriptions by use of the waiting time distribution" (2013) Pharmacoepidemiology and Drug Safety 22(8):803-809. DOI:10.1002/PDS.3459
}
\author{
Claire Deakin
}
