% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_period.R
\name{exposure_period}
\alias{exposure_period}
\title{Function to calculate the time period an individual has been exposed to a dispensed medicine}
\usage{
exposure_period(
  data,
  epe_table,
  id,
  atc_code,
  pbs_code,
  date_dispensing,
  gap_as_epe_multiplier = TRUE,
  gap_days = NULL,
  epe_gap_multiplier = 3,
  epe_gap_multiple_dispensings_same_date = TRUE
)
}
\arguments{
\item{data}{data.frame containing input data; check your input data is a data.frame and not a tibble, coerce to data.frame if necessary. To use, specify data=your_input_data}

\item{epe_table}{data.frame object containing PBS codes and EPE values e.g. if output from `epe` function stored as epe_res, use epe_table = epe_res$epe_table to extract the epe_table from the list object}

\item{id}{name of character vector in `data` specifying a unique identifier for individual people/patients (can handle integer variables). To use, specify id="your_id_variable_name"}

\item{atc_code}{name of character vector in `data` specifying the ATC code for the medicine individuals have been exposed to. Note, if there are multiple ATC codes for a given medicine, you may wish to use the medicine name instead so long as it is a unique identifier for the medicine. To use, specify atc_code="your_atc_code_variable_name"}

\item{pbs_code}{name of character vector in `data` specifying the PBS code for the medicine. To use, specify pbs_code="your_pbs_code_variable_name"}

\item{date_dispensing}{name of date vector in `data` specifying the dates on which each medicine were dispensed. To use, specify date_dispensing="your_date_of_dispensing_variable_name"}

\item{gap_as_epe_multiplier}{argument to specify whether a gap in treatment will be identified based on being an integer value multiplied by the EPE, with that integer value specified in the epe_gap_multiplier argument. Defaults to TRUE}

\item{gap_days}{optional integer value to use if gap_as_epe_multiplier = FALSE, to specify the number of days to use as a gap a gap isn't defined as an integer value multipled by the EPE}

\item{epe_gap_multiplier}{optional argument to specify the integer to multiply the EPE by to define a gap. Defaults to 3 if no value is given by the user}

\item{epe_gap_multiple_dispensings_same_date}{optional argument to specify whether identifying a gap in treatment will accommodate the period covered by multiple dispensings on the same date. Defaults to TRUE i.e. the total period covered by multiple dispensings is accommodated}
}
\value{
Object of class data.frame containing the id, atc_code, exposure_start, exposure_end, number_dispensings and exposure_days
}
\description{
Uses dispensing dates and the output from the `epe` function to calculate the exposure period in days to a given medicine, based on ATC codes and the EPE value calculated for each PBS code.
The exposure period is the number of days from the first date of dispensing until 1x EPE after the last date of dispensing. Gaps in dispensing are excluded from this period based on a default of 3x EPE days, however, other values can be specified e.g. a number of days or a different integer multiplier of the EPE.
Gap periods are calculated based on the EPE value for the PBS code for the medicine dispensed immediately prior to the gap.
}
\examples{
exposure_period(data = your_input_data,
epe_table = epe_res$epe_table,
id = "your_id_variable_name",
atc_code = "your_atc_code_variable_name",
pbs_code = "your_pbs_code_variable_name"
date_dispensing = "your_dispensing_date_variable_name"
)

exposure_period(data = your_input_data,
epe_table = epe_res$epe_table,
id = "your_id_variable_name",
atc_code = "your_atc_code_variable_name",
pbs_code = "your_pbs_code_variable_name",
date_dispensing = "your_dispensing_date_variable_name"
epe_gap_multiplier = 5
)

exposure_period(data = your_input_data,
epe_table = epe_res$epe_table,
id = "your_id_variable_name",
atc_code = "your_atc_code_variable_name",
pbs_code = "your_pbs_code_variable_name",
date_dispensing = "your_dispensing_date_variable_name"
gap_as_epe_multiplier = FALSE,
gap_days = 120
)

Example using 'mydata' sample dataset provided with the package
epe_res <- epe(data=mydata, id="PPN", pbs_code="PBS", date_dispensing="dispensing_date")
exposure_period(data=mydata, epe_table=epe_res$epe_table, id="PPN", atc_code="ATC_code", pbs_code="PBS", date_dispensing="dispensing_date")

}
\references{
Anton Pottegård and Jesper Hallas, "Assigning exposure duration to single prescriptions by use of the waiting time distribution" (2013) Pharmacoepidemiology and Drug Safety 22(8):803-809. DOI:10.1002/PDS.3459
}
\author{
Claire Deakin
}
