% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Made up toy dataset to practise running 'epe' and 'exposure_period' functions}
\format{

}
\usage{
data(mydata)
}
\description{
Simulated data to generate known values for the EPE and exposure period
}
\keyword{datasets}
